memory: add memory handling
This commit adds memory handling for the Netflow exporter arguments.

+++ util.c
@@ -2,7 +2,7 @@
 /*                                                        */
 /* File: util.c                                           */
 /* Created: 2022-10-01                                    */
-/* Last change: 2022-10-01                                */
+/* Last change: 2022-10-13                                */
 /* Author: David Chocholaty <xchoch09@stud.fit.vutbr.cz>  */
 /* Project: Project for the course ISA - variant 1        */
 /*          - Generation of NetFlow data from captured    */
@@ -17,9 +17,11 @@
 #include <stdlib.h>
 
 /*
- * TODO
+ * Function to check if the input string is only from numeric characters.
+ *
+ * @param string Input string for check.
  */
-bool is_numeric_string(char* string)
+bool is_numeric_string (char* string)
 {
     for (char *i = string ; *i ; i++) {
         if (!isdigit(*i))
@@ -34,7 +36,7 @@ bool is_numeric_string(char* string)
 /*
  * Function for convert string to uint16_t.
  *
- * @param string Input string
+ * @param string Input string.
  * @return       If string doesn't contain uint16_t number: 0L,
  *               number in uint16_t data type otherwise.
  */
@@ -53,11 +55,11 @@ uint16_t strtoui_16 (char* string)
 /*
  * Function for convert string to uint32_t.
  *
- * @param string Input string
+ * @param string Input string.
  * @return       If string doesn't contain uint32_t number: 0L,
  *               number in uint32_t data type otherwise.
  */
-uint16_t strtoui_32 (char* string)
+uint32_t strtoui_32 (char* string)
 {
     if (!is_numeric_string(string))
     {
@@ -70,7 +72,12 @@ uint16_t strtoui_32 (char* string)
 }
 
 /*
- * TODO
+ * Function checks if the value is in a range between the minimum
+ * and maximum value including both range sides.
+ *
+ * @param value Input value for check in range.
+ * @param min   Minimum range value.
+ * @param max   Maximum range value.
  */
 bool in_range (const unsigned int value,
                const unsigned int min,
